Split EssentialApp and EssentialFeed configuration into two different files
