+++ Tuist/ProjectDescriptionHelpers/EssentialAppConfiguration.swift
@@ -1,8 +1,5 @@
 import ProjectDescription
 
-
-// MARK: - EssentialApp
-
 public extension Configuration {
     static var essentialAppDebug: Self {
         .debug(
@@ -32,27 +29,5 @@ public extension ConfigurationName {
     
     static var essentialAppRelease: Self {
         .configuration("essentialAppRelease")
-    }
-}
-
-// MARK: - EssentialFeed
-
-public extension Configuration {
-    static var essentialFeedDebug: Self {
-        .debug(name: "Debug", xcconfig: .relativeToRoot("ConfigFiles/EssentialFeed.xcconfig"))
-    }
-    
-    static var essentialFeedRelease: Self {
-        .release(name: "Release", xcconfig: .relativeToRoot("ConfigFiles/EssentialFeed.xcconfig"))
-    }
-}
-
-public extension ConfigurationName {
-    static var essentialFeedDebug: Self {
-        .configuration("essentialFeedDebug")
-    }
-    
-    static var essentialFeedRelease: Self {
-        .configuration("essentialFeedRelease")
     }
 }
